+++ src/main/java/subway/business/domain/direction/LineModifyStrategy.java
@@ -1,8 +1,8 @@
 package subway.business.domain.direction;
 
 import java.util.List;
-import subway.business.domain.Section;
-import subway.business.domain.Station;
+import subway.business.domain.line.Section;
+import subway.business.domain.line.Station;
 
 public interface LineModifyStrategy {
     void addTerminus(Station station, List<Section> sections, int distance);
