+++ src/main/java/subway/business/domain/direction/UpwardLineModifyStrategy.java
@@ -1,8 +1,8 @@
 package subway.business.domain.direction;
 
 import java.util.List;
-import subway.business.domain.Section;
-import subway.business.domain.Station;
+import subway.business.domain.line.Section;
+import subway.business.domain.line.Station;
 
 public class UpwardLineModifyStrategy implements LineModifyStrategy {
