+++ src/main/java/subway/business/domain/line/OrderedStations.java
@@ -1,4 +1,4 @@
-package subway.business.domain;
+package subway.business.domain.line;
 
 import java.util.LinkedList;
 import java.util.List;
