+++ src/main/java/subway/business/service/LineService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.business.domain.Line;
-import subway.business.domain.LineRepository;
 import subway.business.domain.direction.Direction;
+import subway.business.domain.line.Line;
+import subway.business.domain.line.LineRepository;
 import subway.business.service.dto.LineResponse;
 import subway.business.service.dto.LineSaveRequest;
 import subway.business.service.dto.StationAddToLineRequest;
