+++ src/main/java/subway/persistence/repository/DbLineRepository.java
@@ -5,10 +5,10 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
-import subway.business.domain.Line;
-import subway.business.domain.LineRepository;
-import subway.business.domain.Section;
-import subway.business.domain.Station;
+import subway.business.domain.line.Line;
+import subway.business.domain.line.LineRepository;
+import subway.business.domain.line.Section;
+import subway.business.domain.line.Station;
 import subway.persistence.dao.LineDao;
 import subway.persistence.dao.SectionDao;
 import subway.persistence.dao.StationDao;
