+++ src/test/java/subway/business/domain/SectionTest.java
@@ -8,6 +8,8 @@ import static subway.fixture.Fixture.station2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import subway.business.domain.line.Section;
+import subway.business.domain.line.Station;
 
 public class SectionTest {
     @DisplayName("거리를 받아 전체 거리에서 뺀 거리를 계산한다.")
