+++ src/test/java/subway/ui/LineControllerTest.java
@@ -17,9 +17,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import subway.business.domain.Line;
-import subway.business.domain.Section;
-import subway.business.domain.Station;
+import subway.business.domain.line.Line;
+import subway.business.domain.line.Section;
+import subway.business.domain.line.Station;
 import subway.business.service.LineService;
 import subway.business.service.dto.LineResponse;
 import subway.business.service.dto.LineSaveRequest;
